split build API tests to a separate file
